refactor: extract current render validation
Move the cohesive current-render adapter and schema-cloning logic into
a cycle-safe leaf module. Keep nested and decorator orchestration in
the main runtime module.

Preserve the 0.3.0 wire boundary for Field(exclude=True) and exclude_if.
Add a generated-current-wire extra-injection regression for both forms.

Validation: 858 tests passed on Pydantic 2.13.4 (95.82% coverage).
Validation: 858 tests passed on Pydantic 2.12.3 (95.74% coverage).
Validation: Ruff, ty, strict mypy, Vulture, and strict docs passed.
Validation: immutable v0.3.0 contract passed.

Closes #110
Refs #94
